[labenusers] entrega obrigatoria
